python/requirements.txt
influxdb-client>=1.36.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
reportlab>=4.0.0
